refactor helper methods out of models file
