+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/_helpers.py
@@ -4,6 +4,7 @@
 # Licensed under the MIT License.
 # ------------------------------------
 
+import re
 import six
 from azure.core.credentials import AzureKeyCredential
 from azure.core.pipeline.policies import AzureKeyCredentialPolicy
@@ -13,6 +14,7 @@ from azure.core.exceptions import (
     ResourceExistsError,
     ClientAuthenticationError
 )
+import azure.ai.formrecognizer._models as models
 
 POLLING_INTERVAL = 5
 COGNITIVE_KEY_HEADER = "Ocp-Apim-Subscription-Key"
@@ -25,28 +27,86 @@ error_map = {
 }
 
 
-class TransportWrapper(HttpTransport):
-    """Wrapper class that ensures that an inner client created
-    by a `get_client` method does not close the outer transport for the parent
-    when used in a context manager.
+def get_bounding_box(field):
+    return [
+        models.Point(x=field.bounding_box[0], y=field.bounding_box[1]),
+        models.Point(x=field.bounding_box[2], y=field.bounding_box[3]),
+        models.Point(x=field.bounding_box[4], y=field.bounding_box[5]),
+        models.Point(x=field.bounding_box[6], y=field.bounding_box[7])
+    ] if field.bounding_box else None
+
+
+def adjust_value_type(value_type):
+    if value_type == "array":
+        value_type = "list"
+    if value_type == "number":
+        value_type = "float"
+    if value_type == "object":
+        value_type = "dictionary"
+    return value_type
+
+
+def adjust_confidence(score):
+    """Adjust confidence when not returned.
     """
-    def __init__(self, transport):
-        self._transport = transport
+    if score is None:
+        return 1.0
+    return score
 
-    def send(self, request, **kwargs):
-        return self._transport.send(request, **kwargs)
 
-    def open(self):
-        pass
+def adjust_text_angle(text_angle):
+    """Adjust to (-180, 180]
+    """
+    if text_angle > 180:
+        text_angle -= 360
+    return text_angle
 
-    def close(self):
-        pass
 
-    def __enter__(self):
-        pass
+def get_elements(field, read_result):
+    text_elements = []
 
-    def __exit__(self, *args):  # pylint: disable=arguments-differ
-        pass
+    for item in field.elements:
+        nums = [int(s) for s in re.findall(r"\d+", item)]
+        read = nums[0]
+        line = nums[1]
+        if len(nums) == 3:
+            word = nums[2]
+            ocr_word = read_result[read].lines[line].words[word]
+            extracted_word = models.FormWord._from_generated(ocr_word, page=read+1)  # pylint: disable=protected-access
+            text_elements.append(extracted_word)
+            continue
+        ocr_line = read_result[read].lines[line]
+        extracted_line = models.FormLine._from_generated(ocr_line, page=read+1)  # pylint: disable=protected-access
+        text_elements.append(extracted_line)
+    return text_elements
+
+
+def get_field_value(field, value, read_result):  # pylint: disable=too-many-return-statements
+    if value is None:
+        return value
+    if value.type == "string":
+        return value.value_string
+    if value.type == "number":
+        return value.value_number
+    if value.type == "integer":
+        return value.value_integer
+    if value.type == "date":
+        return value.value_date
+    if value.type == "phoneNumber":
+        return value.value_phone_number
+    if value.type == "time":
+        return value.value_time
+    if value.type == "array":
+        return [
+            models.FormField._from_generated(field, value, read_result)  # pylint: disable=protected-access
+            for value in value.value_array
+        ]
+    if value.type == "object":
+        return {
+            key: models.FormField._from_generated(key, value, read_result)  # pylint: disable=protected-access
+            for key, value in value.value_object.items()
+        }
+    return None
 
 
 def get_authentication_policy(credential):
@@ -94,3 +154,27 @@ def check_beginning_bytes(form):
         if form[:4] == b"\x4D\x4D\x00\x2A":  # big-endian
             return "image/tiff"
     raise ValueError("Content type could not be auto-detected. Please pass the content_type keyword argument.")
+
+
+class TransportWrapper(HttpTransport):
+    """Wrapper class that ensures that an inner client created
+    by a `get_client` method does not close the outer transport for the parent
+    when used in a context manager.
+    """
+    def __init__(self, transport):
+        self._transport = transport
+
+    def send(self, request, **kwargs):
+        return self._transport.send(request, **kwargs)
+
+    def open(self):
+        pass
+
+    def close(self):
+        pass
+
+    def __enter__(self):
+        pass
+
+    def __exit__(self, *args):  # pylint: disable=arguments-differ
+        pass
